[ing] 채팅 기능

+++ src/main/java/com/realEstate/realEstate/model/dto/chatting/Message.java
@@ -32,7 +32,7 @@ public class Message implements Serializable {
     private Long senderNo;
 
     @NotNull
-    private Integer saleNo;
+    private Long saleNo;
 
     private long sendTime;
     private Integer readCount;
